Added new Section class and related methods in Morphology.
The new class and Morphology have been placed in a new namespace called
brain::cell.
The new class provides access to morphological data on a per
section basis. From a section it's possible to query the parent and
children as well as information for its sample points. The new
class is very lightweight so it can be created on-the-fly by
brain::Morphology and doesn't need to be stored anywhere.