Change all timestamps in compartment reports to be double.
Fixed all the round-off errorscaused by the new implementation of
timestamp -> framenumber that doesn't use round. The recommended
way of hitting a specific frame now is to ask for the timestamp in
the middle of the frame.

One observation resulting from this is that maybe a function for
requesting a specific frame should be added.