Allow append section whose ID is already present in the morphology
It used to be impossible to append a section whose ID is the same as
another section of the morphology. Now it appends a copy of the section under
a different ID.

Another change is that one can now use set_maximum_warnings(int) to
control the maximum number of warning messages to be displayed on screen.