Implement soma surfaces
Soma surface are computed according to the soma type. SomaContour has not
been implemented yet.

Corrected a bug where the equality check would not pass between the same
morphology written in 2 different file format as this would induce a
different number of soma points in both format. Now the soma points are
excluded from the equality check and the reminder of the points vector is
checked starting at the first non soma points (which is not the same offset
in both format).

Tests data have been moved to a "tests/data" folder