Fix reading SWC with bifurcation points with changing diameters (#44)
* Fix reading SWC with bifurcation points with changing diameters

And create a warning when appending empty sections. This warning used
to be part of the checkDuplicate warning. But this caused the new
reading of SWC files with bifurcation points with changing diameter
to trigger it. So the new warning was created and disabled for the part
of the code that reads SWC.