Add Morphology::sectionIndices, the list of section offsets
In the immutable version, Morphology::points() and Morphology::diameters()
return a list of all points/diameters of the morphology. However it was no very
usable as the information about which point is in which section was not
available. See: https://github.com/BlueBrain/MorphIO/issues/116

This commits adds the method Morphology::sectionOffsets that returns the list
of offset to access the points of a given section.