Implement Tristan's proposal Fix: #148
The soma type was stored in a variable in Soma that was readonly.
When the Neuron was created from scratch the type was set UNDEFINED
(since the neuron wa empty) and there was no way of changing it
afterwards. Now we rely entirely on the heuristic function
morphio::getSomaType() and removed the variable _somaType alltogether.

getSomaType() was also simplified as Tristan suggested: the map
creation&search has been substituted with a switch simplifiing the
code and adding the noexcept operator.

The change was implemented in mut::Soma as well as in Soma. Since
getSomaType() is O(1) it should not degrade performances.

- Remove somaType_ variable from Soma and mut::Soma.
- Simplify morphio::getSomaType adding noexcept for extra security.