🐛 Freeze CellFamily for h5v2
H5v2 format only support 'Neuron' has a CellFamily so we don't
need to read the field and can directly set the value to Neuron.

Moreover, fixed a kind of bug where the metadata from h5v1 in `_readV11Metadata`
would be stored in the morpho properties even if the morpho is not of type h5v1.