Implement API for `load_unordered`.
The point of this API is to enable collections to load morphologies
in any order they see fit. For HDF5 containers tests have shown that the
access pattern matters a lot in terms of performance.

The idea is a generic API that allows optimizing the iteration order for
reorderable loops, e.g.,

    for k, morph_name in enumerate(morphology_names):
        morph = collection.load(morph_name)
        f(k, morph)

can be replaced with

    for k, morph in collection.load_unordered(morphology_names):
        assert collection.load(morphology_names[k]) == morph
        f(k, morph)

This commit only adds the minimum required for this to work. In
particular, it doesn't optimize the access pattern.